+++ byclaw-be/src/main/java/com/iwhalecloud/byai/gateway/channels/service/dingtalk/stream/model/DingtalkMessageDownloadInfo.java
@@ -9,6 +9,7 @@ public class DingtalkMessageDownloadInfo {
 
     private String downloadCode;
     private String fileName;
+    private String downloadUrl;
 
     public DingtalkMessageDownloadInfo(String downloadCode, String fileName) {
         this.downloadCode = downloadCode;
